Added functionality for ExampleProcess and User Process classes and changed overall from int type to double type

+++ java/Customer.java
@@ -13,7 +13,7 @@ public class Customer extends CustomerEvaluation{
     protected double LTV;
     protected double DTI;
     protected double FEDTI;
-    protected double overall;
+    protected int overall;
     
     public String toString()
     {
